+++ src/components/results/results-bitmaptext.js
@@ -8,7 +8,7 @@ import fp from "../../../lib/lodash/fp/fp.js";
 
 export class ResultsBitmapText extends Phaser.GameObjects.BitmapText {
     constructor(scene, config) {
-        super(scene, 0, 0, config.font, undefined, config.size);
+        super(scene, 0, 0, config.bitmapFont, undefined, config.size);
         this.config = config;
         this.setTextFromTemplate(config.content, scene.transientData);
     }